libs/common-sdk/pyproject.toml
[tool.poetry]
name = "tabula-common-sdk"
version = "1.0.0"
description = "Common SDK for environment configuration and utility functions"
packages = [
    { include = "common_sdk" }
]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
pymongo = "^4.13.0"
motor = "^3.7.1"
boto3 = "^1.38.21"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
openai = "^1.81.0"
pytz = "^2025.2"
python-jose = {extras = ["cryptograph"], version = "^3.4.0"}
sse-starlette = {version = "^2.3.5", extras = ["all"]}


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"